refactor: BalenaCloud client to use resource-based structure
- Moved organization, fleet, release, device, and related resource classes into separate files for better organization and maintainability.
- Updated BalenaCloud class to initialize resource clients instead of nested classes.
- Created a new `resources` module to encapsulate all resource-related functionality.
- Removed excessive inline documentation from the BalenaCloud class and moved it to respective resource classes.
